refactor: 패키지 분리 및 ShowRoomInfoListener 이름 변경
