+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roomplayer/RoomInfoView.kt
@@ -5,6 +5,7 @@ import android.view.LayoutInflater
 import androidx.constraintlayout.widget.ConstraintLayout
 import com.digginroom.digginroom.databinding.RoomInfoBinding
 import com.digginroom.digginroom.feature.room.ScrapListener
+import com.digginroom.digginroom.feature.room.customview.roomplayer.navigator.DefaultRoomNavigator
 import com.digginroom.digginroom.model.RoomModel
 
 class RoomInfoView(context: Context) : ConstraintLayout(context) {
@@ -12,9 +13,10 @@ class RoomInfoView(context: Context) : ConstraintLayout(context) {
         RoomInfoBinding.inflate(LayoutInflater.from(context), this, true)
 
     init {
-        roomInfoBinding.showRoomInfoListener = DefaultShowDetailListener(context)
+        roomInfoBinding.showRoomInfoListener = DefaultShowRoomInfoListener(context)
         roomInfoBinding.roomNavigator = DefaultRoomNavigator(context)
     }
+
     fun setRoomInfo(room: RoomModel, onScrapListener: ScrapListener) {
         roomInfoBinding.roomModel = room
         roomInfoBinding.trackModel = room.trackModel
