--- android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roomplayer/RoomNavigator.kt
@@ -1,6 +1,0 @@
-package com.digginroom.digginroom.feature.room.customview.roomplayer
-
-interface RoomNavigator {
-
-    fun navigateToScrap()
-}
