+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roomplayer/ShowRoomInfoListener.kt
@@ -2,7 +2,7 @@ package com.digginroom.digginroom.feature.room.customview.roomplayer
 
 import com.digginroom.digginroom.model.TrackModel
 
-interface ShowDetailListener {
+interface ShowRoomInfoListener {
 
     fun show(trackModel: TrackModel)
 }
