+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roomplayer/navigator/RoomNavigator.kt
@@ -1,0 +1,6 @@
+package com.digginroom.digginroom.feature.room.customview.roomplayer.navigator
+
+interface RoomNavigator {
+
+    fun navigateToScrap()
+}
